refactor: split cashier.py + telegram_bot_enhanced.py + billing_service.py
== cashier.py (1787 LOC → 5 modules) ==
- _helpers.py: imports, helpers, Pydantic models, router
- _payments.py (8 endpoints): pending-payments, payments CRUD, grouped, cancel, confirm, refund
- _stats.py (3 endpoints): stats, export, hourly
- _visits.py (2 endpoints): mark-paid, receipt

== telegram_bot_enhanced.py (1668 LOC → 7 modules) ==
- _base.py: imports, header, EnhancedTelegramBotServiceMixinBase
- _core.py (7 methods): __init__, _handle_command, _check_admin_rights, _send_*, send_*_notification
- _admin.py (10 methods): _handle_admin_* handlers
- _patient.py (6 methods): _handle_start, _handle_menu, _handle_queue, etc.
- _appointments.py (4 methods): _handle_appointments, _handle_book, _handle_cancel, _handle_reschedule
- _info.py (6 methods): _handle_profile, _handle_doctors, _handle_services, etc.
- __init__.py: composes all mixins + global instance + get_enhanced_telegram_bot

== billing_service.py (1527 LOC → 8 modules) ==
- _base.py: imports, header, BillingServiceMixinBase
- _core.py (6 methods): __init__, _get_local_timestamp_naive, _normalize_datetime, _generate_*
- _invoices.py (5 methods): create_invoice, auto_generate_*, generate_invoice_html, send_invoice
- _payments.py (6 methods): create_payment, get_payments_list, is_visit_paid, validate_*, update_*, record_*
- _calculations.py (6 methods): calculate_total, get_discount_mode_for_visit, etc.
- _reminders.py (5 methods): create_payment_reminders, send_due_reminders, etc.
- _recurring.py (3 methods): create_recurring_invoices, _calculate_next_recurrence_date, get_billing_settings
- __init__.py: composes all mixins

All public APIs preserved. 34 tests pass. 0 route duplicates. 1128 routes.
